requirements.txt
# FastAPI dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==0.8.0
websockets==12.0
starlette==0.27.0
pydantic==2.5.2

# AI and image processing
torch==2.1.0
transformers>=4.0.0
Pillow==10.1.0
safetensors==0.4.0
accelerate==0.4.0
sentencepiece==0.1.99

# Other dependencies
python-dotenv>=0.19.0
tqdm==4.67.1
aiohappyeyeballs==2.4.4
aiosignal==1.3.2
annotated-types==0.7.0
attrdict==2.0.1
attrs==25.1.0
av==14.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
einops==0.8.0
exceptiongroup==1.2.0
filelock==3.17.0
fonttools==4.55.8
frozenlist==1.5.0
fsspec==2025.2.0
ftfy==6.3.1
huggingface-hub==0.28.1
httptools>=0.5.0
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.5
kiwisolver==1.4.8
lightning-utilities==0.12.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy>=1.19.0
open_clip_torch==2.30.0
packaging==23.2
pandas==2.2.3
pluggy==1.5.0
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
pytest==8.3.4
python-dateutil==2.9.0
pytorch-lightning==2.5.0.post0
pytz==2025.1
PyYAML==6.0.2
qwen-vl-utils==0.0.8
regex==2024.11.6
requests==2.32.3
six==1.16.0
sympy==1.13.1
timm==1.0.14
tokenizers==0.21.0
torchmetrics==1.6.1
torchvision>=0.15.0
typing_extensions==4.10.0
tzdata==2025.1
urllib3==2.3.0
uvloop>=0.17.0
yarl==1.18.3